+++ xnlib/xnacrylic/include/font.h
@@ -15,7 +15,7 @@
 #include <canvas.h>
 
 typedef struct _font_ {
-	uint8_t *data;
+	uint64_t *data;
 	char *name;
 	int size;
 	uint32_t width;
